requirements.txt
google-generativeai
flask
Gunicorn